+++ tests/test_endpoints.py
@@ -4,6 +4,7 @@ import pytest
 from aioresponses.core import aioresponses
 
 from keba_keenergy_api.api import KebaKeEnergyAPI
+from keba_keenergy_api.constants import ExternalHeatSourceHeatRequest
 from keba_keenergy_api.constants import ExternalHeatSourceOperatingMode
 from keba_keenergy_api.constants import HeatCircuitExternalCoolRequest
 from keba_keenergy_api.constants import HeatCircuitExternalHeatRequest
@@ -2357,6 +2358,107 @@ class TestHeatPumpSection:
                 ssl=False,
             )
 
+    @pytest.mark.asyncio
+    async def test_get_operating_time(self) -> None:
+        """Test get operating time."""
+        with aioresponses() as mock_keenergy_api:
+            mock_keenergy_api.post(
+                "http://mocked-host/var/readWriteVars",
+                payload=[
+                    {
+                        "name": "APPL.CtrlAppl.sParam.heatpump[0].operationalData.operationalTimeS",
+                        "attributes": {
+                            "formatId": "fmt6p0",
+                            "longText": "Operational hrs.",
+                            "unitId": "TimeHour",
+                        },
+                        "value": "65037217",
+                    }
+                ],
+                headers={"Content-Type": "application/json;charset=utf-8"},
+            )
+
+            client: KebaKeEnergyAPI = KebaKeEnergyAPI(host="mocked-host")
+            data: int = await client.heat_pump.get_operating_time()
+
+            assert isinstance(data, int)
+            assert data == 65037217  # noqa: PLR2004
+
+            mock_keenergy_api.assert_called_once_with(
+                url="http://mocked-host/var/readWriteVars",
+                data='[{"name": "APPL.CtrlAppl.sParam.heatpump[0].operationalData.operationalTimeS", "attr": "1"}]',
+                method="POST",
+                auth=None,
+                ssl=False,
+            )
+
+    @pytest.mark.asyncio
+    async def test_get_max_runtime(self) -> None:
+        """Test get max runtime."""
+        with aioresponses() as mock_keenergy_api:
+            mock_keenergy_api.post(
+                "http://mocked-host/var/readWriteVars",
+                payload=[
+                    {
+                        "name": "APPL.CtrlAppl.sParam.heatpump[0].operationalData.maxRunTimeS",
+                        "attributes": {
+                            "formatId": "fmt6p1",
+                            "longText": "Max run-time",
+                            "unitId": "TimeHour",
+                        },
+                        "value": "8836597",
+                    }
+                ],
+                headers={"Content-Type": "application/json;charset=utf-8"},
+            )
+
+            client: KebaKeEnergyAPI = KebaKeEnergyAPI(host="mocked-host")
+            data: int = await client.heat_pump.get_max_runtime()
+
+            assert isinstance(data, int)
+            assert data == 8836597  # noqa: PLR2004
+
+            mock_keenergy_api.assert_called_once_with(
+                url="http://mocked-host/var/readWriteVars",
+                data='[{"name": "APPL.CtrlAppl.sParam.heatpump[0].operationalData.maxRunTimeS", "attr": "1"}]',
+                method="POST",
+                auth=None,
+                ssl=False,
+            )
+
+    @pytest.mark.asyncio
+    async def test_get_activation_counter(self) -> None:
+        """Test get activation counter."""
+        with aioresponses() as mock_keenergy_api:
+            mock_keenergy_api.post(
+                "http://mocked-host/var/readWriteVars",
+                payload=[
+                    {
+                        "name": "APPL.CtrlAppl.sParam.heatpump[0].operationalData.activationCounter",
+                        "attributes": {
+                            "formatId": "fmt6p0",
+                            "longText": "Turn-on cycles",
+                        },
+                        "value": "1197",
+                    }
+                ],
+                headers={"Content-Type": "application/json;charset=utf-8"},
+            )
+
+            client: KebaKeEnergyAPI = KebaKeEnergyAPI(host="mocked-host")
+            data: int = await client.heat_pump.get_activation_counter()
+
+            assert isinstance(data, int)
+            assert data == 1197  # noqa: PLR2004
+
+            mock_keenergy_api.assert_called_once_with(
+                url="http://mocked-host/var/readWriteVars",
+                data='[{"name": "APPL.CtrlAppl.sParam.heatpump[0].operationalData.activationCounter", "attr": "1"}]',
+                method="POST",
+                auth=None,
+                ssl=False,
+            )
+
 
 class TestHeatCircuitSection:
     @pytest.mark.asyncio
@@ -3527,6 +3629,153 @@ class TestExternalHeatSourceSection:
             mock_keenergy_api.assert_called_once_with(
                 url="http://mocked-host/var/readWriteVars",
                 data='[{"name": "APPL.CtrlAppl.sParam.extHeatSource[0].values.setTemp", "attr": "1"}]',
+                method="POST",
+                auth=None,
+                ssl=False,
+            )
+
+    @pytest.mark.asyncio
+    @pytest.mark.parametrize(
+        ("human_readable", "payload_value", "expected_value"),
+        [
+            (True, "true", "on"),
+            (False, ExternalHeatSourceHeatRequest.ON.value, 1),
+            (True, "false", "off"),
+            (False, ExternalHeatSourceHeatRequest.OFF.value, 0),
+        ],
+    )
+    async def test_get_heat_request(
+        self,
+        human_readable: bool,  # noqa: FBT001
+        payload_value: str,
+        expected_value: str,
+    ) -> None:
+        """Test get heat request."""
+        with aioresponses() as mock_keenergy_api:
+            mock_keenergy_api.post(
+                "http://mocked-host/var/readWriteVars",
+                payload=[
+                    {
+                        "name": "APPL.CtrlAppl.sParam.extHeatSource[0].DO.values.setValueB",
+                        "attributes": {
+                            "longText": "Dig. request",
+                        },
+                        "value": payload_value,
+                    }
+                ],
+                headers={"Content-Type": "application/json;charset=utf-8"},
+            )
+
+            client: KebaKeEnergyAPI = KebaKeEnergyAPI(host="mocked-host")
+            data: int | str = await client.external_heat_source.get_heat_request(human_readable=human_readable)
+
+            assert isinstance(data, (int | str))
+            assert data == expected_value
+
+            mock_keenergy_api.assert_called_once_with(
+                url="http://mocked-host/var/readWriteVars",
+                data='[{"name": "APPL.CtrlAppl.sParam.extHeatSource[0].DO.values.setValueB", "attr": "1"}]',
+                method="POST",
+                auth=None,
+                ssl=False,
+            )
+
+    @pytest.mark.asyncio
+    async def test_get_operating_time(self) -> None:
+        """Test get operating time."""
+        with aioresponses() as mock_keenergy_api:
+            mock_keenergy_api.post(
+                "http://mocked-host/var/readWriteVars",
+                payload=[
+                    {
+                        "name": "APPL.CtrlAppl.sParam.extHeatSource[0].DO.operationalData.operationalTimeS",
+                        "attributes": {
+                            "formatId": "fmt6p0",
+                            "longText": "Operational hrs.",
+                            "unitId": "TimeHour",
+                        },
+                        "value": "3809028",
+                    }
+                ],
+                headers={"Content-Type": "application/json;charset=utf-8"},
+            )
+
+            client: KebaKeEnergyAPI = KebaKeEnergyAPI(host="mocked-host")
+            data: int = await client.external_heat_source.get_operating_time()
+
+            assert isinstance(data, int)
+            assert data == 3809028  # noqa: PLR2004
+
+            mock_keenergy_api.assert_called_once_with(
+                url="http://mocked-host/var/readWriteVars",
+                data='[{"name": "APPL.CtrlAppl.sParam.extHeatSource[0].DO.operationalData.operationalTimeS", "attr": "1"}]',
+                method="POST",
+                auth=None,
+                ssl=False,
+            )
+
+    @pytest.mark.asyncio
+    async def test_get_max_runtime(self) -> None:
+        """Test get max runtime."""
+        with aioresponses() as mock_keenergy_api:
+            mock_keenergy_api.post(
+                "http://mocked-host/var/readWriteVars",
+                payload=[
+                    {
+                        "name": "APPL.CtrlAppl.sParam.extHeatSource[0].DO.operationalData.maxRunTimeS",
+                        "attributes": {
+                            "formatId": "fmt6p1",
+                            "longText": "Max run-time",
+                            "unitId": "TimeHour",
+                        },
+                        "value": "602403",
+                    }
+                ],
+                headers={"Content-Type": "application/json;charset=utf-8"},
+            )
+
+            client: KebaKeEnergyAPI = KebaKeEnergyAPI(host="mocked-host")
+            data: int = await client.external_heat_source.get_max_runtime()
+
+            assert isinstance(data, int)
+            assert data == 602403  # noqa: PLR2004
+
+            mock_keenergy_api.assert_called_once_with(
+                url="http://mocked-host/var/readWriteVars",
+                data='[{"name": "APPL.CtrlAppl.sParam.extHeatSource[0].DO.operationalData.maxRunTimeS", "attr": "1"}]',
+                method="POST",
+                auth=None,
+                ssl=False,
+            )
+
+    @pytest.mark.asyncio
+    async def test_get_activation_counter(self) -> None:
+        """Test get activation counter."""
+        with aioresponses() as mock_keenergy_api:
+            mock_keenergy_api.post(
+                "http://mocked-host/var/readWriteVars",
+                payload=[
+                    {
+                        "name": "APPL.CtrlAppl.sParam.extHeatSource[0].DO.operationalData.activationCounter",
+                        "attributes": {
+                            "formatId": "fmt6p0",
+                            "longText": "Turn-on cycles",
+                        },
+                        "value": "477",
+                    }
+                ],
+                headers={"Content-Type": "application/json;charset=utf-8"},
+            )
+
+            client: KebaKeEnergyAPI = KebaKeEnergyAPI(host="mocked-host")
+            data: int = await client.external_heat_source.get_activation_counter()
+
+            assert isinstance(data, int)
+            assert data == 477  # noqa: PLR2004
+
+            mock_keenergy_api.assert_called_once_with(
+                url="http://mocked-host/var/readWriteVars",
+                data='[{"name": "APPL.CtrlAppl.sParam.extHeatSource[0].DO.operationalData.activationCounter", "attr": "1"}]',
                 method="POST",
                 auth=None,
                 ssl=False,